backend/requirements.txt
flask
requests
psycopg2-binary
sentence-transformers
faiss-cpu